Mejoras en la barra de navegación y aplicadas optimizaciones más agresivas en el wasm
