+++ docs/index.html
@@ -4,21 +4,18 @@
     <head>
         <meta charset="utf-8" />
         <title>sfmolina portfolio</title>
-        <link rel="stylesheet" href="/styles-d8eb2b559b26c318.css" integrity="sha384&#x2D;OZQVhwxb970bN1JgmQnYGlhJ6X2XGKN6W2AFbVGmZa78&#x2B;F4NVmfDBVZk&#x2F;W5I7hjV"/>
+        <link rel="stylesheet" href="/styles-7eab6b79b99a6ea1.css" integrity="sha384&#x2D;&#x2B;qJjHjH2wKPjBI3KuXHyAZRO7LlwhLYDkDb5rJ3p7gn7fhvmnvSNj9Z6caeBJmXq"/>
         
         <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
-    
-<link rel="modulepreload" href="/sfmolina-personal-page-97c64ed5ea85ea72.js" crossorigin=anonymous integrity="sha384-iCtMQMEJaUOKHzNiZhO+GtUJpyabXvJ1z+Nf+afa4NdWgc0k0ApHAUGkVLGmHO7q">
-<link rel="modulepreload" href="/snippets/sfmolina-personal-page-8932d4c58286d8d9/inline0.js" crossorigin=anonymous integrity="sha384-/r4JY5FmMSaflEjDt0EaBxWjnCljyEUqCWkPJM7+IEUO6A3YrGalAmDnLXR8/v+d">
-<link rel="preload" href="/sfmolina-personal-page-97c64ed5ea85ea72_bg.wasm" crossorigin=anonymous integrity="sha384-bloYY/5rIiWVTVodc9wRF00SGZSb9iHOM/RgRZLmxa3UBxSlKqnqN6di6B8DKwqj" as="fetch" type="application/wasm"></head>
+    <link rel="modulepreload" href="/sfmolina-personal-page-39c578ac5e30b2d0.js" crossorigin=anonymous integrity="sha384-fXZwzTzsRyS842085L+AXK5oIx9XAjX3u0Ol132Yed/DQLO1H8ToOc4aIkdzfpdn"><link rel="modulepreload" href="/snippets/sfmolina-personal-page-8932d4c58286d8d9/inline0.js" crossorigin=anonymous integrity="sha384-/r4JY5FmMSaflEjDt0EaBxWjnCljyEUqCWkPJM7+IEUO6A3YrGalAmDnLXR8/v+d"><link rel="preload" href="/sfmolina-personal-page-39c578ac5e30b2d0_bg.wasm" crossorigin=anonymous integrity="sha384-pWFg6r9h2bTAhTSwyY4AR5FUMCSQvGFs9b1KkNl+FAjgnXyns9phZnWgKt1nfbRA" as="fetch" type="application/wasm"></head>
 
     <body>
         <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
         <script src=" https://cdn.jsdelivr.net/npm/echarts@5.5.1/dist/echarts.min.js "></script>
     
-<script type="module" nonce="k+7e+1oo0Anr6zPwWegKVg==">
-import init, * as bindings from '/sfmolina-personal-page-97c64ed5ea85ea72.js';
-const wasm = await init('/sfmolina-personal-page-97c64ed5ea85ea72_bg.wasm');
+<script type="module" nonce="QvMoo8awXlH+CFfXnaHO/g==">
+import init, * as bindings from '/sfmolina-personal-page-39c578ac5e30b2d0.js';
+const wasm = await init('/sfmolina-personal-page-39c578ac5e30b2d0_bg.wasm');
 
 
 window.wasmBindings = bindings;
