+++ sfmolina-personal-page/src/app.rs
@@ -9,9 +9,9 @@
 use yew::prelude::*;
 use yew_router::prelude::*;
 use crate::{
-    global_ctx::GlobalContextProvider,
-    router::{switch, Route},
-    components::nav_bar::Navbar,
+    components::nav_bar::Navbar, 
+    global_ctx::GlobalContextProvider, 
+    router::{switch, Route}
 };
