+++ sfmolina-personal-page/src/components/nav_bar.rs
@@ -1,7 +1,7 @@
 //-------------------------------------------------------------------//
 //  AUTHOR:    @sfmolina                                            //
 //  Version:   v1                                                  //
-//  Modified:  23oc24                                             //
+//  Modified:  19de24                                             //
 //---------------------------------------------------------------//
 
 
@@ -9,10 +9,8 @@
 use yew::prelude::*;
 use yew_router::prelude::*;
 use crate::{
-    router::Route,
-    global_ctx::GlobalContext,
-    translations::{
-        ENGLISH_TEXTS, SPANISH_TEXTS, LanguageVariant
+    global_ctx::GlobalContext, router::Route, translations::{
+        LanguageVariant, ENGLISH_TEXTS, SPANISH_TEXTS
     }
 };
 
@@ -20,7 +18,6 @@ use crate::{
 
 #[function_component(Navbar)]
 pub fn navbar() -> Html {
-
     let global_ctx = use_context::<GlobalContext>().expect("GlobalContext not found");
     let language = global_ctx.language;
 
@@ -35,130 +32,118 @@ pub fn navbar() -> Html {
         })
     };
 
-    let current_route: Option<Route> = use_route();
-
     html! {
-        <nav class="navbar navbar-expand-lg navbar-light bg-transparent fixed-top">
+        <div class="nav-style">
+        <nav class="navbar navbar-expand-lg bg-body-tertiary fixed-top">
             <div class="container-fluid">
-                <Link<Route> classes={classes!("navbar-brand")} to={Route::Root}>
-                    {language.author_username}
-                </Link<Route>>
-                <button 
-                    class={
-                        match current_route {
-                            Some(Route::AboutMe) => {
-                                "navbar-toggler"
-                            }
-                            _ => {
-                                "navbar-toggler white"
-                            }
-                        }
-                    } 
-                    type="button" 
-                    data-bs-toggle="collapse" 
-                    data-bs-target="#navbarNav" 
-                    aria-controls="navbarNav" 
-                    aria-expanded="false" 
-                    aria-label="Toggle navigation"
-                >
+                // Navbar brand
+                <Link<Route> classes={classes!("navbar-brand")} to={Route::Root}>{language.author_username}</Link<Route>>
+
+                // Navbar toggler (for small screens)
+                <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" 
+                        aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
                     <span class="navbar-toggler-icon"></span>
                 </button>
-                <div class="collapse navbar-collapse" id="navbarNav">
-                    <ul class="navbar-nav me-auto">
-                        <li class="nav-item">
-                            <Link<Route> classes={
-                                match current_route {
-                                    Some(Route::AboutMe) => {
-                                        classes!("nav-link", "activated")
-                                    }
-                                    _ => {
-                                        classes!("nav-link")
-                                    }
-                                }
-                            } to={Route::AboutMe}>
-                                {language.nav.about_me}
-                            </Link<Route>>
-                        </li>
-                        <li class="nav-item">
-                            <Link<Route> classes={
-                                match current_route {
-                                    Some(Route::Projects) => {
-                                        classes!("nav-link", "activated")
-                                    }
-                                    _ => {
-                                        classes!("nav-link")
-                                    }
-                                }
-                            } to={Route::Projects}>
-                                {language.nav.projects}
-                            </Link<Route>>
-                        </li>
-                        <li class="nav-item dropdown">
-                            <button 
-                                class="nav-link dropdown-toggle projects-color"
-                                id="navbarDropdown" 
-                                data-bs-toggle="dropdown" 
-                                aria-expanded="false"
-                            >
-                                {language.nav.misc}
-                            </button>
-                            <ul class="dropdown-menu dropdown-menu-end" aria-labelledby="navbarDropdown">
-                                <li>
-                                    <a 
-                                        class={"nav-link"} 
-                                        href="https://sfmolina.github.io/supertres/"
-                                    >
-                                        {language.nav.super_tres}
-                                    </a>
-                                </li>
-                                <li>
-                                    <a 
-                                        class={"nav-link"} 
-                                        href="https://sfmolina.github.io/mod-comp/"
-                                    >
-                                        {language.nav.mod_comp}
-                                    </a>
-                                </li>
-                            </ul>
-                        </li>
-                    </ul>
-                    <ul class="navbar-nav">
-                        <li class="nav-item dropdown">
-                            <button 
-                                class={
-                                    match current_route {
-                                        Some(Route::AboutMe) => {"nav-link dropdown-toggle default-color"},
-                                        _ => {"nav-link dropdown-toggle projects-color"}
-                                    }
-                                } 
-                                id="navbarDropdown" 
-                                data-bs-toggle="dropdown" 
-                                aria-expanded="false"
-                            >
-                                {language.nav.language}
-                            </button>
-                            <ul class="dropdown-menu dropdown-menu-end" aria-labelledby="navbarDropdown">
-                                <li>
-                                    <button 
-                                        class="dropdown-item" 
-                                        onclick={on_language_change.reform(|_| LanguageVariant::English)}
-                                    >
-                                        {language.nav.en}
-                                    </button>
-                                </li>
-                                <li>
-                                    <button 
-                                        class="dropdown-item" 
-                                        onclick={on_language_change.reform(|_| LanguageVariant::Spanish)}
-                                    >
-                                        {language.nav.es}
-                                    </button>
-                                </li>
-                            </ul>
-                        </li>
-                    </ul>
+
+                // Navbar itself
+                <div class="collapse navbar-collapse" id="navbarSupportedContent">
+                    // Navbar left side
+                    <LefttNav />
+
+                    // Navbar right side
+                    <RightNav on_language_change={on_language_change} />
                 </div>
             </div>
         </nav>
+        </div>
     }
+}
+
+
+#[function_component(LefttNav)]
+fn left_nav() -> Html {
+
+    let global_ctx = use_context::<GlobalContext>().expect("GlobalContext not found");
+    let language = global_ctx.language;
+    let current_route: Option<Route> = use_route();
+
+    html! {
+        <ul class="navbar-nav me-auto mb-2 mb-lg-0">
+                        
+            // Link to AboutMe
+            <li class="nav-item">
+                <Link<Route> classes={classes!(
+                    "nav-link",
+                    (current_route == Some(Route::AboutMe)).then_some("active")
+                )} to={Route::AboutMe}>{language.nav.about_me}</Link<Route>>
+            </li>
+            
+            // Link to Projects
+            <li class="nav-item">
+                <Link<Route> classes={classes!(
+                    "nav-link",
+                    (current_route == Some(Route::Projects)).then_some("active")
+                )} to={Route::Projects}>{language.nav.projects}</Link<Route>>
+            </li>
+            
+            // Links to other projects
+            <li class="nav-item dropdown">
+                <a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
+                    {language.nav.misc}
+                </a>
+                <ul class="dropdown-menu">
+                    <li><a class="dropdown-item" href="https://sfmolina.github.io/supertres/">{language.nav.super_tres}</a></li>
+                    
+                    <li><hr class="dropdown-divider"/></li>
+                    
+                    <li><a class="dropdown-item" href="https://sfmolina.github.io/mod-comp/">{language.nav.mod_comp}</a></li>
+                </ul>
+            </li>
+    
+        </ul>
+    }
+
+}
+
+
+#[derive(Properties, PartialEq)]
+struct RightNavProps {
+    on_language_change: Callback<LanguageVariant>,
+}
+#[function_component(RightNav)]
+fn right_nav(RightNavProps { on_language_change }: &RightNavProps) -> Html {
+
+    let global_ctx = use_context::<GlobalContext>().expect("GlobalContext not found");
+    let language = global_ctx.language;
+
+    html! {
+        <ul class="navbar-nav">
+            <li class="nav-item dropdown">
+                <a class="nav-link dropdown-toggle bold" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
+                    {language.nav.language}
+                </a>
+                <ul class="dropdown-menu dropdown-menu-end">
+                    <li>
+                        <button 
+                            class="dropdown-item" 
+                            onclick={on_language_change.reform(|_| LanguageVariant::English)}
+                        >
+                            {language.nav.en}
+                        </button>
+                    </li>
+                    <li>
+                        <button 
+                            class="dropdown-item" 
+                            onclick={on_language_change.reform(|_| LanguageVariant::Spanish)}
+                        >
+                            {language.nav.es}
+                        </button>
+                    </li>
+                    //<li><hr class="dropdown-divider"/></li>
+                    //<li><a class="dropdown-item" href="#">{"Something else here"}</a></li>
+                </ul>
+            </li>
+        </ul>
+    }
+
 }
